+++ app/main.rb
@@ -1,2 +1,4 @@
 # require 'app/basic_sample.rb'
-require 'app/interactive_sample.rb'
+require 'app/basic_sample_v3.rb'
+# require 'app/interactive_sample.rb'
+# require 'app/interactive_sample_v3.rb'